requirements.txt
python-telegram-bot
sqlalchemy
flask
gunicorn
pysqlite
python-dotenv
